+++ graphs/src/anemoi/graphs/edges/builders/multi_scale.py
@@ -40,6 +40,10 @@ class MultiScaleEdges(BaseEdgeBuilder):
     x_hops : int
         Number of hops (in the refined icosahedron) between two nodes to connect
         them with an edge.
+    scale_resolutions : Union[int, List[int], None]
+        Defines the refinement levels at which edges are computed. If an integer is provided, edges are computed for all
+        levels up to and including that level. For instance, `scale_resolutions=4` includes edges at levels 1 through 4,
+        whereas `scale_resolutions=[4]` only includes edges at level 4.
 
     Methods
     -------
@@ -55,12 +59,22 @@ class MultiScaleEdges(BaseEdgeBuilder):
     `MultiScaleEdges` only supports computing the edges within a set of nodes built by an `Type[IcosahedronNodes]`.
     """
 
-    def __init__(self, source_name: str, target_name: str, x_hops: int, **kwargs):
+    def __init__(
+        self, source_name: str, target_name: str, x_hops: int, scale_resolutions: int | list[int] | None, **kwargs
+    ):
         super().__init__(source_name, target_name)
         assert source_name == target_name, f"{self.__class__.__name__} requires source and target nodes to be the same."
         assert isinstance(x_hops, int), "Number of x_hops must be an integer"
         assert x_hops > 0, "Number of x_hops must be positive"
         self.x_hops = x_hops
+        if isinstance(scale_resolutions, int):
+            assert scale_resolutions > 0, "The scale_resolutions argument only supports positive integers."
+            scale_resolutions = list(range(1, scale_resolutions + 1))
+        assert not isinstance(scale_resolutions, str), "The scale_resolutions argument is not valid."
+        assert (
+            scale_resolutions is None or min(scale_resolutions) > 0
+        ), "The scale_resolutions argument only supports positive integers."
+        self.scale_resolutions = scale_resolutions
 
     @staticmethod
     def get_edge_builder_class(node_type: str) -> Type[BaseIcosahedronEdgeStrategy]:
@@ -80,7 +94,16 @@ class MultiScaleEdges(BaseEdgeBuilder):
 
     def compute_edge_index(self, source_nodes: NodeStorage, _target_nodes: NodeStorage) -> torch.Tensor:
         edge_builder_cls = MultiScaleEdges.get_edge_builder_class(source_nodes.node_type)
-        source_nodes = edge_builder_cls().add_edges(source_nodes, self.x_hops)
+
+        # Get the refinement levels (or scales) at which to compute the neighbourhoods.
+        scale_resolutions = self.scale_resolutions or source_nodes["_resolutions"]
+        if self.scale_resolutions is not None and max(self.scale_resolutions) > max(source_nodes["_resolutions"]):
+            LOGGER.warning(
+                f"Some scale resolutions may be ignored because they are greater than the resolution of the nodes ({max(source_nodes['_resolutions'])})."
+            )
+
+        # Add edges
+        source_nodes = edge_builder_cls().add_edges(source_nodes, self.x_hops, scale_resolutions=scale_resolutions)
         adjmat = nx.to_scipy_sparse_array(source_nodes["_nx_graph"], format="coo")
 
         # Get source & target indices of the edges