+++ graphs/src/anemoi/graphs/schemas/node_attributes_schemas.py
@@ -16,7 +16,7 @@ from typing import Union
 
 from pydantic import Field
 
-from anemoi.training.schemas.utils import BaseModel
+from anemoi.utils.schemas import BaseModel
 
 LOGGER = logging.getLogger(__name__)
 
@@ -28,6 +28,15 @@ class PlanarAreaWeightSchema(BaseModel):
         "anemoi.graphs.nodes.attributes.UniformWeights",
     ] = Field(..., alias="_target_")
     "Implementation of the area of the nodes as the weights from anemoi.graphs.nodes.attributes."
+    norm: Literal["unit-max", "l1", "l2", "unit-sum", "unit-std"] = Field(example="unit-max")
+    "Normalisation of the weights."
+
+
+class MaskedPlanarAreaWeightsSchema(BaseModel):
+    target_: Literal["anemoi.graphs.nodes.attributes.MaskedPlanarAreaWeights"] = Field(..., alias="_target_")
+    "Implementation of the area of the nodes as the weights from anemoi.graphs.nodes.attributes."
+    mask_node_attr_name: str = Field(examples="cutout_mask")
+    "Attribute name to mask the area weights."
     norm: Literal["unit-max", "l1", "l2", "unit-sum", "unit-std"] = Field(example="unit-max")
     "Normalisation of the weights."
 
@@ -58,6 +67,7 @@ class NonmissingAnemoiDatasetVariableSchema(BaseModel):
 
 SingleAttributeSchema = Union[
     PlanarAreaWeightSchema,
+    MaskedPlanarAreaWeightsSchema,
     SphericalAreaWeightSchema,
     CutOutMaskSchema,
     NonmissingAnemoiDatasetVariableSchema,