+++ graphs/tests/conftest.py
@@ -61,6 +61,9 @@ def graph_with_nodes() -> HeteroData:
     graph["test_nodes"].x = 2 * torch.pi * torch.tensor(coords)
     graph["test_nodes"].mask = torch.tensor([True] * len(coords)).unsqueeze(-1)
     graph["test_nodes"].mask2 = torch.tensor([True] * (len(coords) - 2) + [False] * 2).unsqueeze(-1)
+    graph["test_nodes"].interior_mask = torch.tensor(
+        [False, False, False, False, False, True, True, False, False, False, False, False]
+    ).unsqueeze(-1)
     graph["test_nodes"]["_grid_reference_distance"] = 0.75
     return graph
 
@@ -81,7 +84,20 @@ def graph_nodes_and_edges() -> HeteroData:
     graph = HeteroData()
     graph["test_nodes"].x = 2 * torch.pi * torch.tensor(coords)
     graph["test_nodes"].mask = torch.tensor([True] * len(coords)).unsqueeze(-1)
-    graph[("test_nodes", "to", "test_nodes")].edge_index = torch.tensor([[0, 1, 2, 3], [1, 2, 3, 0]])
+    graph[("test_nodes", "to", "test_nodes")].edge_index = torch.tensor([[3, 1, 2, 0], [2, 0, 1, 3]])
+    graph[("test_nodes", "to", "test_nodes")].edge_attr = (
+        10 * graph[("test_nodes", "to", "test_nodes")].edge_index[0][:, None]
+    )
+    return graph
+
+
+@pytest.fixture
+def graph_long_and_short_edges() -> HeteroData:
+    """Graph with a pair of short (800km) and a pair of long (20000km) edges."""
+    graph = HeteroData()
+    graph["test_nodes"].x = 2 * torch.pi * torch.tensor([[-0.01, 0], [0.01, 0], [-0.01, 0.5], [0.01, 0.5]])
+    graph["test_nodes"]["southern_hemisphere_mask"] = torch.tensor([[1], [0], [1], [0]], dtype=torch.bool)
+    graph["test_nodes", "to", "test_nodes"].edge_index = torch.tensor([[0, 0, 1, 3], [1, 3, 2, 2]])
     return graph