+++ training/src/anemoi/training/data/datamodule/singledatamodule.py
@@ -69,6 +69,10 @@ class AnemoiDatasetsDataModule(pl.LightningDataModule):
         return self.ds_train.statistics
 
     @cached_property
+    def statistics_tendencies(self) -> dict:
+        return self.ds_train.statistics_tendencies
+
+    @cached_property
     def metadata(self) -> dict:
         return self.ds_train.metadata
 
@@ -214,6 +218,7 @@ class AnemoiDatasetsDataModule(pl.LightningDataModule):
         return NativeGridDataset(
             data_reader=data_reader,
             relative_date_indices=self.relative_date_indices(val_rollout),
+            timestep=self.config.data.timestep,
             shuffle=shuffle,
             grid_indices=self.grid_indices,
             label=label,