+++ training/src/anemoi/training/diagnostics/mlflow/logger.py
@@ -42,6 +42,8 @@ if TYPE_CHECKING:
 
 LOGGER = logging.getLogger(__name__)
 
+MAX_PARAMS_LENGTH = 2000
+
 
 class LogsMonitor:
     """Class for logging terminal output.
@@ -305,6 +307,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
 
         self._fork_run_server2server = None
         self._parent_run_server2server = None
+        self._parent_dry_run = False
 
         enabled = authentication and not offline
         self.auth = TokenAuth(tracking_uri, enabled=enabled)
@@ -345,6 +348,16 @@ class AnemoiMLflowLogger(MLFlowLogger):
             prefix=prefix,
             run_id=run_id,
         )
+
+    def _check_dry_run(self, run: mlflow.entities.Run) -> None:
+        """Check if the parent run is a dry run.
+
+        A dry run is a run that is used as template base run
+        but do not contain any checkpoints.
+        """
+        dry_run = run.data.tags.get("dry_run", "False") == "True"
+        LOGGER.info("Parent run is a Dry Run: %s", dry_run)
+        self._parent_dry_run = dry_run
 
     def _check_server2server_lineage(self, run: mlflow.entities.Run) -> bool:
         """Address lineage and metadata for server2server runs.
@@ -398,6 +411,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
                 parent_run = mlflow_client.get_run(parent_run_id)
                 run_name = parent_run.info.run_name
                 self._check_server2server_lineage(parent_run)
+                self._check_dry_run(parent_run)
                 tags["mlflow.parentRunId"] = parent_run_id
                 tags["resumedRun"] = "True"  # tags can't take boolean values
             # This block is used when a run ID is specified without child runs option activated
@@ -406,6 +420,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
                 run = mlflow_client.get_run(run_id)
                 run_name = run.info.run_name
                 self._check_server2server_lineage(run)
+                self._check_dry_run(parent_run)
                 mlflow_client.update_run(run_id=run_id, status="RUNNING")
                 tags["resumedRun"] = "True"
             # This block is used when a run is forked and an existing run ID is specified
@@ -415,6 +430,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
                 parent_run = mlflow_client.get_run(parent_run_id)
                 run_name = parent_run.info.run_name
                 self._check_server2server_lineage(parent_run)
+                self._check_dry_run(parent_run)
                 tags["mlflow.parentRunId"] = config_run_id  # We want to be linked to the main run ID
                 tags["resumedRun"] = "True"  # We want to be linked to the main run ID
                 tags["forkedRun"] = "True"  # This is a forked run
@@ -426,6 +442,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
                 tags["forkedRunId"] = parent_run_id
                 run = mlflow_client.get_run(parent_run_id)
                 self._check_server2server_lineage(run)
+                self._check_dry_run(run)
 
         if not run_name:
             import uuid
@@ -600,6 +617,11 @@ class AnemoiMLflowLogger(MLFlowLogger):
             )  # Flatten dict with '.' to not break API queries
             if clean_params:
                 expanded_params = clean_config_params(expanded_params)
+
+            LOGGER.info("Logging %s parameters", len(expanded_params))
+
+            if len(expanded_params) > MAX_PARAMS_LENGTH:
+                LOGGER.warning("Logging a large number of parameters to %s", len(expanded_params))
 
             # Truncate parameter values.
             params_list = [Param(key=k, value=str(v)[:truncation_length]) for k, v in expanded_params.items()]